Refactor core to load after game start
